rename nsq.async to nsq.conn due to python3.7 keyword conflict
async keyword was added to python3.5, made strict in python3.7

leave locally unused async.py with reference to AsyncConn for
backwards-compatibility with older code that imports nsq.async.AsyncConn

add python3.7 to travis-ci test matrix
