+++ .travis.yml
@@ -32,6 +32,11 @@ matrix:
     dist: xenial
     sudo: true
 
+  - python: "3.7"
+    env: TORNADO_VERSION=4.5.3 NSQ_DOWNLOAD=nsq-1.0.0-compat.linux-amd64.go1.8
+    dist: xenial
+    sudo: true
+
 script:
   - ./travis_test.sh
 notifications:
