+++ nsq/writer.py
@@ -9,8 +9,8 @@ import random
 from ._compat import string_types
 from ._compat import func_args
 from .client import Client
-from nsq import protocol
-from . import async
+from .conn import AsyncConn
+from . import protocol
 
 logger = logging.getLogger(__name__)
 
@@ -96,7 +96,7 @@ class Writer(Client):
         self.conns = {}
 
         # Verify keyword arguments
-        valid_args = func_args(async.AsyncConn.__init__)
+        valid_args = func_args(AsyncConn.__init__)
         diff = set(kwargs) - set(valid_args)
         assert len(diff) == 0, 'Invalid keyword argument(s): %s' % list(diff)
 
@@ -172,7 +172,7 @@ class Writer(Client):
         assert isinstance(host, string_types)
         assert isinstance(port, int)
 
-        conn = async.AsyncConn(host, port, **self.conn_kwargs)
+        conn = AsyncConn(host, port, **self.conn_kwargs)
         conn.on('identify', self._on_connection_identify)
         conn.on('identify_response', self._on_connection_identify_response)
         conn.on('auth', self._on_connection_auth)
